# Speed up getindex with AbstractVector{Bool} row selection